+++ frontend/src/page/exercises.rs
@@ -17,6 +17,7 @@ pub fn init(mut url: Url, orders: &mut impl Orders<Msg>, navbar: &mut crate::Nav
     navbar.title = String::from("Exercises");
 
     Model {
+        search_term: String::new(),
         dialog: Dialog::Hidden,
         loading: false,
     }
@@ -27,6 +28,7 @@ pub fn init(mut url: Url, orders: &mut impl Orders<Msg>, navbar: &mut crate::Nav
 // ------ ------
 
 pub struct Model {
+    search_term: String,
     dialog: Dialog,
     loading: bool,
 }
@@ -53,6 +55,7 @@ pub enum Msg {
     ShowDeleteExerciseDialog(usize),
     CloseExerciseDialog,
 
+    SearchTermChanged(String),
     NameChanged(String),
 
     SaveExercise,
@@ -89,6 +92,9 @@ pub fn update(
             Url::go_and_replace(&crate::Urls::new(&data_model.base_url).exercises());
         }
 
+        Msg::SearchTermChanged(search_term) => {
+            model.search_term = search_term;
+        }
         Msg::NameChanged(name) => match model.dialog {
             Dialog::AddExercise(ref mut form) | Dialog::EditExercise(ref mut form) => {
                 if data_model.exercises.iter().all(|e| e.name != name) {
@@ -144,7 +150,11 @@ pub fn update(
 pub fn view(model: &Model, data_model: &data::Model) -> Node<Msg> {
     div![
         view_exercise_dialog(&model.dialog, &data_model.exercises, model.loading),
-        view_table(data_model),
+        div![
+            C!["px-4"],
+            common::view_search_box(&model.search_term, Msg::SearchTermChanged)
+        ],
+        view_table(&model.search_term, data_model),
         common::view_fab("plus", |_| Msg::ShowAddExerciseDialog),
     ]
 }
@@ -236,7 +246,7 @@ fn view_exercise_dialog(dialog: &Dialog, exercises: &[data::Exercise], loading: 
     )
 }
 
-fn view_table(data_model: &data::Model) -> Node<Msg> {
+fn view_table(search_term: &str, data_model: &data::Model) -> Node<Msg> {
     div![
         C!["table-container"],
         C!["mt-4"],
@@ -244,15 +254,17 @@ fn view_table(data_model: &data::Model) -> Node<Msg> {
             C!["table"],
             C!["is-fullwidth"],
             C!["is-hoverable"],
-            thead![tr![th!["Name"], th![]]],
             tbody![&data_model
                 .exercises
                 .iter()
                 .enumerate()
+                .filter(|(_, e)| e.name.to_lowercase().contains(&search_term.to_lowercase()))
                 .map(|(i, e)| {
                     let id = e.id;
-                    tr![
-                        td![a![
+                    tr![td![
+                        C!["is-flex"],
+                        C!["is-justify-content-space-between"],
+                        a![
                             attrs! {
                                 At::Href => {
                                     crate::Urls::new(&data_model.base_url)
@@ -261,8 +273,8 @@ fn view_table(data_model: &data::Model) -> Node<Msg> {
                                 }
                             },
                             e.name.to_string(),
-                        ]],
-                        td![p![
+                        ],
+                        p![
                             C!["is-flex is-flex-wrap-nowrap"],
                             a![
                                 C!["icon"],
@@ -276,8 +288,8 @@ fn view_table(data_model: &data::Model) -> Node<Msg> {
                                 ev(Ev::Click, move |_| Msg::ShowDeleteExerciseDialog(i)),
                                 i![C!["fas fa-times"]]
                             ]
-                        ]]
-                    ]
+                        ]
+                    ]]
                 })
                 .collect::<Vec<_>>()],
         ]